Add CustomStringConvertible and Equatable conformances to Position and LSPRange
