+++ SwiftLSPClient/Types/Basic.swift
@@ -10,26 +10,6 @@ import Foundation
 
 public typealias DocumentUri = String
 
-public struct Position: Codable {
-    public let line: Int
-    public let character: Int
-    
-    public init(line: Int, character: Int) {
-        self.line = line
-        self.character = character
-    }
-}
-
-public struct LSPRange: Codable {
-    public let start: Position
-    public let end: Position
-    
-    public init(start: Position, end: Position) {
-        self.start = start
-        self.end = end
-    }
-}
-
 public enum MarkupKind: String, Codable {
     case plaintext
     case markdown
